refactor: merge share functions.ts and functionsNG.ts files
