+++ packages/web-pkg/tests/unit/composables/actions/files/useFileActionsCopyQuicklink.spec.ts
@@ -9,7 +9,7 @@ import { FileAction } from '../../../../../src/composables/actions'
 import { useCanShare } from '../../../../../src/composables/shares'
 import { Resource, SpaceResource } from '@ownclouders/web-client'
 import { LinkShare } from '@ownclouders/web-client/src/helpers/share'
-import { buildLinkShare } from '@ownclouders/web-client/src/helpers/share/functionsNG'
+import { buildLinkShare } from '@ownclouders/web-client/src/helpers/share/functions'
 import { useClipboard } from '../../../../../src/composables/clipboard'
 import { useMessages } from '../../../../../src/composables/piniaStores'
 import { Permission } from '@ownclouders/web-client/src/generated'
@@ -26,7 +26,8 @@ vi.mock('../../../../../src/composables/clipboard', () => ({
   useClipboard: vi.fn()
 }))
 
-vi.mock('@ownclouders/web-client/src/helpers/share/functionsNG', () => ({
+vi.mock('@ownclouders/web-client/src/helpers/share/functions', async (importOriginal) => ({
+  ...(await importOriginal<any>()),
   buildLinkShare: vi.fn()
 }))
